requirements.txt
streamlit
google-generativeai
python-dotenv
crewai
rich
